Refactor pytest service module and update imports
- Introduce a new pytest.py module for processing pytest results, encapsulating related functionality.
- Update import paths in existing modules to reflect the new structure, ensuring consistency across the codebase.
- Modify test files to align with the new pytest module, maintaining test integrity.
