+++ src/mcp_suite/servers/qa/service/__init__.py
@@ -6,6 +6,6 @@
 from mcp_suite.servers.qa import logger
 
 from .coverage import process_coverage_json
-from .pytest_service import process_pytest_results
+from .pytest import process_pytest_results
 
 __all__ = ["process_coverage_json", "process_pytest_results", "logger"]
