+++ src/mcp_suite/servers/qa/service/tests/test_pytest_service.py
@@ -1,4 +1,4 @@
-"""Tests for the pytest_service module."""
+"""Tests for the pytest module."""
 
 import json
 from unittest.mock import mock_open, patch
@@ -7,7 +7,7 @@ from mcp_suite.servers.qa.config import ReportPaths
 from mcp_suite.servers.qa.models.pytest_models import (
     PytestResults,
 )
-from mcp_suite.servers.qa.service.pytest_service import (
+from mcp_suite.servers.qa.service.pytest import (
     process_pytest_results,
 )
