+++ src/mcp_suite/servers/qa/tools/pytest_tool.py
@@ -18,7 +18,7 @@ import time
 from pathlib import Path
 
 from mcp_suite.servers.qa import logger as main_logger
-from mcp_suite.servers.qa.service.pytest_service import process_pytest_results
+from mcp_suite.servers.qa.service.pytest import process_pytest_results
 from mcp_suite.servers.qa.utils.decorators import exception_handler
 from mcp_suite.servers.qa.utils.git_utils import get_git_root
 from mcp_suite.servers.qa.utils.logging_utils import get_component_logger
